+++ dart/lib/src/protocol/dsn.dart
@@ -9,11 +9,9 @@ class Dsn {
   });
 
   /// The Sentry.io public key for the project.
-  @visibleForTesting
   final String publicKey;
 
   /// The Sentry.io secret key for the project.
-  @visibleForTesting
   final String secretKey;
 
   /// The ID issued by Sentry.io to your project.
@@ -23,6 +21,26 @@ class Dsn {
 
   /// The DSN URI.
   final Uri uri;
+
+  String get postUri {
+    final port = uri.hasPort &&
+            ((uri.scheme == 'http' && uri.port != 80) ||
+                (uri.scheme == 'https' && uri.port != 443))
+        ? ':${uri.port}'
+        : '';
+
+    final pathLength = uri.pathSegments.length;
+
+    String apiPath;
+    if (pathLength > 1) {
+      // some paths would present before the projectID in the uri
+      apiPath =
+          (uri.pathSegments.sublist(0, pathLength - 1) + ['api']).join('/');
+    } else {
+      apiPath = 'api';
+    }
+    return '${uri.scheme}://${uri.host}$port/$apiPath/$projectId/store/';
+  }
 
   static Dsn parse(String dsn) {
     final uri = Uri.parse(dsn);