+++ dart/lib/src/transport/noop_transport.dart
@@ -1,0 +1,11 @@
+import 'dart:async';
+
+import 'package:sentry/sentry.dart';
+
+class NoOpTransport implements Transport {
+  @override
+  Dsn get dsn => null;
+
+  @override
+  Future<SentryId> send(SentryEvent event) => Future.value(SentryId.empty());
+}