+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 - Ref: Sentry init with null and empty DSN and close method #126
 - Ref: Hint is passed across Sentry static class, Hub and Client #124
 - Ref: Remove stackFrameFilter in favor of beforeSendCallback #125
+- Ref: added Transport #123
+- Feat: apply sample rate
 
 # `package:sentry` changelog